Move constants to common namespace
