+++ grapher/Models/Calculations/AccelCalculator.cs
@@ -1,4 +1,5 @@
-﻿using grapher.Models.Serialized;
+﻿using grapher.Common;
+using grapher.Models.Serialized;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
