+++ grapher/Models/Calculations/Data/AccelDataCombined.cs
@@ -1,4 +1,5 @@
-﻿using grapher.Models.Charts;
+﻿using grapher.Common;
+using grapher.Models.Charts;
 using System;
 using System.Collections.Generic;
 using System.Linq;
