+++ grapher/Models/Charts/ChartState/ChartState.cs
@@ -1,4 +1,5 @@
-﻿using grapher.Models.Calculations;
+﻿using grapher.Common;
+using grapher.Models.Calculations;
 using grapher.Models.Calculations.Data;
 using System;
 using System.Windows.Forms;
