+++ grapher/Models/Options/ApplyOptions.cs
@@ -1,5 +1,5 @@
-﻿using grapher.Models.Options.Directionality;
-using grapher.Models.Serialized;
+﻿using grapher.Common;
+using grapher.Models.Options.Directionality;
 using System;
 using System.Drawing;
 using System.Windows.Forms;
