+++ grapher/Models/Options/Directionality/DirectionalityOptions.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using grapher.Common;
 using grapher.Models.Theming;
 
 namespace grapher.Models.Options.Directionality
