+++ grapher/Models/Options/SwitchOption.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using grapher.Common;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
