+++ grapher/Models/Serialized/GUISettings.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using grapher.Common;
+using Newtonsoft.Json;
 using System;
 using System.ComponentModel;
 using System.IO;
